AI-TtS/requirements.txt
pytesseract
Pillow
gTTS
nltk
transformers
torch
pdf2image
pyttsx3
